+++ fixtures/tutorial/app/entry.client.tsx
@@ -1,0 +1,8 @@
+import ReactDOM from "react-dom";
+import Remix from "@remix-run/react/browser";
+
+ReactDOM.hydrate(
+  // @ts-ignore
+  <Remix />,
+  document
+);