+++ fixtures/tutorial/app/entry.server.tsx
@@ -2,8 +2,6 @@ import ReactDOMServer from "react-dom/server";
 import type { EntryContext } from "@remix-run/core";
 import Remix from "@remix-run/react/server";
 
-import App from "./App";
-
 export default function handleRequest(
   request: Request,
   responseStatusCode: number,
@@ -11,9 +9,7 @@ export default function handleRequest(
   remixContext: EntryContext
 ) {
   let markup = ReactDOMServer.renderToString(
-    <Remix context={remixContext} url={request.url}>
-      <App />
-    </Remix>
+    <Remix context={remixContext} url={request.url} />
   );
 
   return new Response("<!DOCTYPE html>" + markup, {