+++ fixtures/tutorial/app/global-data.ts
@@ -1,5 +1,0 @@
-import type { Loader } from "@remix-run/data";
-
-export let loader: Loader = async () => {
-  return { date: new Date() };
-};