+++ CHANGES.md
@@ -2,7 +2,20 @@
 
 This is a history of changes to [Remix](https://remix.run).
 
-## Unreleased
+## 0.13.3 - Wed Feb 24 2021
+
+### Bug Fixes
+
+- Fixed ignoring `*.client.js` and `*.server.js` files in server/browser builds (respectively) when the `.js` extension is missing from the `import`
+- Allow dependencies (`node_modules`) to have side effects
+
+## 0.13.2 - Tue Feb 23 2021
+
+### Bug Fixes
+
+- Fixed manual catchall (\*) routes
+
+## 0.13.1 - Wed Feb 17 2021
 
 ### Improvements